requirements.txt
httpx>=0.27.0
openai>=1.30.0
rich>=13.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pyyaml>=6.0.0
